refactor: 추천 상품 페이지에서 추가한 상품 뒤로 가기시 데이터 동기화
